Added window for creation of a new task - CreateTaskWindow - opens by clicking the Create button on the MainWindow.
Added logic for creating a new task in the form CreateTaskWindow.

Added logic for deleting a task using the Delete button on the MainWindow.

Added ChechNAssignService class for checking and assigning values obtained from the form to a new instance of the Task.

Added HelpService class for adding other auxiliary methods to it.